chatbot/requirements.txt
flask
requests
python-dotenv
certifi
numpy
sentence-transformers
faiss-cpu